reorganize views/projects/ subfolders more
