+++ moped-editor/src/views/projects/newProjectView/NewProjectView.js
